Fixed changes
