+++ .gitignore
@@ -1,8 +1,7 @@
 *.iml
 .gradle
+/.idea
 /local.properties
-/.idea/workspace.xml
-/.idea/libraries
 .DS_Store
 /build
 /captures
