+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,14 @@
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".DataActivity"
+            android:label="@string/list_data">
+            </activity>
     </application>
 
 </manifest>
