--- retrofit-2.3.0/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('retrofit-2.3.0.jar'))
